+++ src/lenders/compound/interface/IOracle.sol
@@ -1,5 +1,5 @@
 pragma solidity >=0.5.8 <0.8.0;
 
 interface IOracle {
-    function getUnderlyingPrice(address) external view returns (uint);
+    function getUnderlyingPrice(address) external view returns (uint256);
 }
